新增 OrdersManagementController 的 findAll 和 insert 功能

+++ src/main/java/com/willy/malltest/repository/UsersRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
 @Component
-public interface UsersRepository extends JpaRepository<User, Long> {
+public interface UsersRepository extends JpaRepository<User, Integer> {
 
 }
